requirements.txt
# ============================================================
# Multimodal RAG - Requirements
# Tested on: RTX 3050 4GB | CUDA 12.1 | Python 3.10/3.11
# LLM: qwen2.5:1.5b via Ollama (runs on CPU, no GPU needed)
# ============================================================

# --- UI ---
streamlit==1.32.0

# --- Document Processing ---
PyPDF2==3.0.1
python-docx==1.1.0

# --- Image Processing ---
Pillow==10.2.0

# --- Audio / Speech-to-Text ---
# Whisper runs on CPU to save GPU VRAM for embeddings
openai-whisper==20231117

# --- ML Stack (CUDA 12.1 compatible, conflict-free) ---
# Install PyTorch separately with the right CUDA wheel:
#   pip install torch==2.2.2+cu121 --index-url https://download.pytorch.org/whl/cu121
torch==2.2.2
transformers==4.38.2
tokenizers==0.15.2
huggingface-hub==0.21.4
sentence-transformers==2.6.1

# --- Vector Database ---
chromadb==0.4.24

# --- Utilities ---
numpy==1.26.4
requests==2.31.0
